Agentpolis version raised, Guice assisted inject version raised, get%velocity methods reverted back to double,
Former-commit-id: 695037edb264632f51018789014896f20758fd60

+++ src/test/java/cz/agents/amodsim/ridesharing/vga/mock/TestOnDemandVehicle.java
@@ -47,8 +47,8 @@ public class TestOnDemandVehicle implements MovingEntity{
 	
 
 	@Override
-	public int getVelocity() {
-		return config.vehicleSpeedInMeters;
+	public double getVelocity() {
+		return (double) config.vehicleSpeedInMeters;
 	}
 
 	@Override
